+++ lib/Phpfastcache/Core/Item/ItemExtendedTrait.php
@@ -360,5 +360,4 @@ trait ItemExtendedTrait
 
         return $info;
     }
-
 }