+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Added (required) cookie driver option `awareOfUntrustableData` to enforce developer awareness of non-reliable data storage
 - The "ActOnAll Helper" have been removed in profit of aggregated cluster support
 - Implemented #713 // Reworked "tags" feature by adding 3 strategies: `TAG_STRATEGY_ONE`, `TAG_STRATEGY_ALL`, `TAG_STRATEGY_ONLY`
+- Removed *global static* properties `CacheManager::$ReadHits`, `CacheManager::$WriteHits` replaced by`\Phpfastcache\Entities\DriverIO` callable in`\Phpfastcache\Core\Pool\ExtendedCacheItemPoolInterface::getIO`